clean up, prepare for testing
