+++ src/server/graphql/mutations/updateReflectionGroupTitle.js
@@ -3,10 +3,9 @@ import getRethink from 'server/database/rethinkDriver';
 import {getUserId, isTeamMember} from 'server/utils/authorization';
 import {sendTeamAccessError} from 'server/utils/authorizationErrors';
 import {sendReflectionGroupNotFoundError} from 'server/utils/docNotFoundErrors';
-import {sendAlreadyCompletedMeetingPhaseError, sendAlreadyEndedMeetingError} from 'server/utils/alreadyMutatedErrors';
+import {sendAlreadyEndedMeetingError} from 'server/utils/alreadyMutatedErrors';
 import publish from 'server/utils/publish';
-import {GROUP, TEAM} from 'universal/utils/constants';
-import isPhaseComplete from 'universal/utils/meetings/isPhaseComplete';
+import {TEAM} from 'universal/utils/constants';
 import stringSimilarity from 'string-similarity';
 import sendSegmentEvent from 'server/utils/sendSegmentEvent';
 import {sendGroupTitleDuplicateError, sendGroupTitleRequiredError} from 'server/utils/__tests__/validationErrors';
